Moved OBJ Loader into render module
