+++ src/shared/pyramid.hpp
